refactor(#1250): move client-shim import rewrite into adapters
`rewriteTemplateImports` lived in the compiler and decided how to handle
`@barefootjs/client(/runtime)` imports (rewrite to a shim source, or
strip) based on the adapter's `clientShimSource`. That was a leak: the
compiler was making per-adapter decisions about adapter output.

Move the helper out as `rewriteImportsForTemplate` in the adapters
package; adapters that re-emit imports (Hono, test) call it themselves.
Adapters that never carry imports in their templates (Go, Mojo) only
read `metadata.templateImports` for diagnostics like BF103, so they
don't need the helper at all.

`buildMetadata` no longer takes a `clientShimSource` argument; it just
hands the raw `imports` list through as `templateImports` for adapter
consumption.
