+++ packages/jsx/src/adapters/test-adapter.ts
@@ -19,6 +19,7 @@ import type {
 } from '../types'
 import type { AdapterOutput, TemplateSections } from './interface'
 import { type JsxAdapterConfig, JsxAdapter } from './jsx-adapter'
+import { rewriteImportsForTemplate } from './template-imports'
 
 export class TestAdapter extends JsxAdapter {
   name = 'test'
@@ -58,8 +59,12 @@ export class TestAdapter extends JsxAdapter {
   private generateImports(ir: ComponentIR): string {
     const lines: string[] = []
 
-    // Use templateImports (client-side packages already filtered by compiler)
-    for (const imp of ir.metadata.templateImports) {
+    // TestAdapter has no client shim, so client-side packages are dropped.
+    const templateImports = rewriteImportsForTemplate(
+      ir.metadata.templateImports,
+      undefined,
+    )
+    for (const imp of templateImports) {
       if (imp.specifiers.length === 0) {
         if (!imp.isTypeOnly) {
           lines.push(`import '${imp.source}'`)
